requirements.txt
pandas
numpy
matplotlib
seaborn
plotly.express